--- Breeze.UI/src/app/wallet/menu/dashboard.component.html
@@ -1,2 +1,0 @@
-<p>Balance: {{confirmedBalance | coinNotation | coinAbbreviation}}</p>
-<p>Unconfirmed: {{unconfirmedBalance | coinNotation | coinAbbreviation}}<p>
