--- Breeze.UI/src/app/wallet/wallet.component.css
@@ -1,12 +1,0 @@
-#sidepanel {
-    margin: 0;
-    padding: 0;
-    height: 100%;
-    width: 200px;
-    position: fixed;
-    overflow: auto;
-}
-
-#content {
-    margin-left: 200px;
-}
