+++ Breeze.UI/src/app/wallet/wallet.component.html
@@ -1,5 +1,5 @@
 <div id="container">
-  <div id="sidepanel">
+  <div id="navigation">
     <app-menu></app-menu>
   </div>
   <div id="content">
