+++ Breeze.UI/src/app/wallet/wallet.module.ts
@@ -4,7 +4,7 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { WalletComponent } from './wallet.component';
 import { MenuComponent } from './menu/menu.component';
-import { DashboardComponent } from './menu/dashboard.component';
+import { DashboardComponent } from './dashboard/dashboard.component';
 import { SendComponent } from './send/send.component';
 import { ReceiveComponent } from './receive/receive.component';
 import { HistoryComponent } from './history/history.component';
